support lint for sub-package usage
The given NOT to import package will be linted not only for itself, but
also for its sub-packages. The lint message will also reflect this.

A potential counterpart change would be to merge the input argument and
merge them if they has root-sub relationship. Or we should allow user to
pass in two paths, e.g. *a=x* and *a/b=y*, and lint the import path of
*a/b* using the second hint (i.e. use suggestion of *y*), this allows
package author to split a mono package into several independent
packages.